Crate mem6_common

mem6_common

version: 19.10.21-20.7

commons for mem6 wasm and server
Learning to code Rust for a http + WebSocket.
Here are just the structures, that are in common between frontend and backend.
Mostly because of the Messages.

Structs

Player
data for one player

Enums

GameStatus
the game can be in various statuses and that differentiate the UI and actions all players have the same game status

MsgAckKind
msg ack kind

WsMessage
WsMessage enum for WebSocket

Constants

_IMPL_DESERIALIZE_FOR_GameStatus

_IMPL_DESERIALIZE_FOR_MsgAckKind

_IMPL_DESERIALIZE_FOR_Player

_IMPL_DESERIALIZE_FOR_WsMessage

_IMPL_SERIALIZE_FOR_GameStatus

_IMPL_SERIALIZE_FOR_MsgAckKind

_IMPL_SERIALIZE_FOR_Player

_IMPL_SERIALIZE_FOR_WsMessage