Enum mem6_common::WsMessage

pub enum WsMessage {
    MsgDummy {
        dummy: String,
    },
    MsgRequestWsUid {
        my_ws_uid: usize,
        players_ws_uid: String,
    },
    MsgResponseWsUid {
        your_ws_uid: usize,
        server_version: String,
    },
    MsgInvite {
        my_ws_uid: usize,
        my_nickname: String,
        asked_game_name: String,
    },
    MsgAccept {
        my_ws_uid: usize,
        players_ws_uid: String,
        my_nickname: String,
    },
    MsgStartGame {
        my_ws_uid: usize,
        players_ws_uid: String,
        players: String,
        card_grid_data: String,
        game_config: String,
    },
    MsgClick1stCard {
        my_ws_uid: usize,
        players_ws_uid: String,
        card_index_of_first_click: usize,
        msg_id: usize,
    },
    MsgClick2ndCardPoint {
        my_ws_uid: usize,
        players_ws_uid: String,
        card_index_of_second_click: usize,
        msg_id: usize,
    },
    MsgTakeTurnBegin {
        my_ws_uid: usize,
        players_ws_uid: String,
        card_index_of_second_click: usize,
        msg_id: usize,
    },
    MsgGameOver {
        my_ws_uid: usize,
        players_ws_uid: String,
    },
    MsgTakeTurnEnd {
        my_ws_uid: usize,
        players_ws_uid: String,
        msg_id: usize,
    },
    MsgAllGameData {
        my_ws_uid: usize,
        players_ws_uid: String,
        players: String,
        card_grid_data: String,
        card_index_of_first_click: usize,
        card_index_of_second_click: usize,
        player_turn: usize,
        game_status: GameStatus,
    },
    MsgAck {
        my_ws_uid: usize,
        players_ws_uid: String,
        msg_id: usize,
        msg_ack_kind: MsgAckKind,
    },
}

WsMessage enum for WebSocket

Variants
MsgDummy

MsgDummy

Fields of MsgDummy

dummy: String

anything

MsgRequestWsUid

Request WebSocket Uid - first message to WebSocket server

Fields of MsgRequestWsUid

my_ws_uid: usize

ws client instance unique id. To not listen the echo to yourself.

players_ws_uid: String

json of vector of players for the server to know whom to send msg

MsgResponseWsUid

response from WebSocket server for first message

Fields of MsgResponseWsUid

your_ws_uid: usize

WebSocket Uid

server_version: String

server version

MsgInvite

invite

Fields of MsgInvite

my_ws_uid: usize

ws client instance unique id. To not listen the echo to yourself.

my_nickname: String

my nickname

asked_game_name: String

content folder name

MsgAccept

accept invitation

Fields of MsgAccept

my_ws_uid: usize

ws client instance unique id. To not listen the echo to yourself.

players_ws_uid: String

json of vector of players for the server to know whom to send msg

my_nickname: String

my nickname

MsgStartGame

player1 initialize the game data and sends it to all players I will send json string to not confuse the server with vectors

Fields of MsgStartGame

my_ws_uid: usize

ws client instance unique id. To not listen the echo to yourself.

players_ws_uid: String

json of vector of players for the server to know whom to send msg

players: String

json of vector of players with nicknames and order data

card_grid_data: String

vector of cards status

game_config: String

json of game_config

MsgClick1stCard

player click

Fields of MsgClick1stCard

my_ws_uid: usize

this identifies the smartphone, but not the player-in-turn

players_ws_uid: String

all players for the server to know whom to send msg

card_index_of_first_click: usize

have to send all the state of the game

msg_id: usize

msg id (random)

MsgClick2ndCardPoint

player click success

Fields of MsgClick2ndCardPoint

my_ws_uid: usize

this identifies the smartphone, but not the player-in-turn

players_ws_uid: String

all players for the server to know whom to send msg

card_index_of_second_click: usize

have to send all the state of the game

msg_id: usize

msg id (random)

MsgTakeTurnBegin

take turn begin

Fields of MsgTakeTurnBegin

my_ws_uid: usize

this identifies the smartphone, but not the player-in-turn

players_ws_uid: String

all players for the server to know whom to send msg

card_index_of_second_click: usize

have to send all the state of the game

msg_id: usize

msg id (random)

MsgGameOver

Play Again

Fields of MsgGameOver

my_ws_uid: usize

this identifies the smartphone, but not the player-in-turn

players_ws_uid: String

all players for the server to know whom to send msg

MsgTakeTurnEnd

player change

Fields of MsgTakeTurnEnd

my_ws_uid: usize

ws client instance unique id. To not listen the echo to yourself.

players_ws_uid: String

all players for the server to know whom to send msg

msg_id: usize

msg id (random)

MsgAllGameData

all game data

Fields of MsgAllGameData

my_ws_uid: usize

ws client instance unique id. To not listen the echo to yourself.

players_ws_uid: String

only the players that recconected

players: String

json of vector of players with nicknames and order data

card_grid_data: String

vector of cards status

card_index_of_first_click: usize

have to send all the state of the game

card_index_of_second_click: usize

have to send all the state of the game

player_turn: usize

player turn

game_status: GameStatus

game status

MsgAck

acknowledge msg, that the receiver received the message

Fields of MsgAck

my_ws_uid: usize

msg sender uid

players_ws_uid: String

send msg to this players

msg_id: usize

msg id (random)

msg_ack_kind: MsgAckKind

kind of ack msg