Enum mem6_common::WsMessage

pub enum WsMessage {
    MsgDummy {
        dummy: String,
    },
    MsgRequestWsUid {
        my_ws_uid: usize,
        players_ws_uid: String,
    },
    MsgResponseWsUid {
        your_ws_uid: usize,
        server_version: String,
    },
    MsgPing {
        msg_id: u32,
    },
    MsgPong {
        msg_id: u32,
    },
    MsgJoin {
        my_ws_uid: usize,
        players_ws_uid: String,
        my_nickname: String,
    },
    MsgStartGame {
        my_ws_uid: usize,
        players_ws_uid: String,
        players: String,
        card_grid_data: String,
        game_config: String,
        game_name: String,
    },
    MsgClick1stCard {
        my_ws_uid: usize,
        players_ws_uid: String,
        card_index_of_first_click: usize,
        msg_id: usize,
    },
    MsgClick2ndCard {
        my_ws_uid: usize,
        players_ws_uid: String,
        card_index_of_second_click: usize,
        is_point: bool,
        msg_id: usize,
    },
    MsgDrinkEnd {
        my_ws_uid: usize,
        players_ws_uid: String,
    },
    MsgGameOver {
        my_ws_uid: usize,
        players_ws_uid: String,
    },
    MsgPlayAgain {
        my_ws_uid: usize,
        players_ws_uid: String,
    },
    MsgTakeTurn {
        my_ws_uid: usize,
        players_ws_uid: String,
        msg_id: usize,
    },
    MsgAck {
        my_ws_uid: usize,
        players_ws_uid: String,
        msg_id: usize,
        msg_ack_kind: MsgAckKind,
    },
    MsgAskPlayer1ForResync {
        my_ws_uid: usize,
        players_ws_uid: String,
    },
    MsgAllGameData {
        my_ws_uid: usize,
        players_ws_uid: String,
        players: String,
        card_grid_data: String,
        card_index_of_first_click: usize,
        card_index_of_second_click: usize,
        player_turn: usize,
        game_status: GameStatus,
    },
}

WsMessage enum for WebSocket

Variants

MsgDummy

MsgDummy

Fields of MsgDummy

dummy: String

anything

MsgRequestWsUid

Request WebSocket Uid - first message to WebSocket server

Fields of MsgRequestWsUid

my_ws_uid: usize

ws client instance unique id. To not listen the echo to yourself.

players_ws_uid: String

json of vector of players for the server to know whom to send msg

MsgResponseWsUid

response from WebSocket server for first message

Fields of MsgResponseWsUid

your_ws_uid: usize

WebSocket Uid

server_version: String

server version

MsgPing

MsgPing

Fields of MsgPing

msg_id: u32

random msg_id

MsgPong

MsgPong

Fields of MsgPong

msg_id: u32

random msg_id

MsgJoin

join the group

Fields of MsgJoin

my_ws_uid: usize

ws client instance unique id. To not listen the echo to yourself.

players_ws_uid: String

json of vector of players for the server to know whom to send msg

my_nickname: String

my nickname

MsgStartGame

player1 initialize the game data and sends it to all players I will send json string to not confuse the server with vectors

Fields of MsgStartGame

my_ws_uid: usize

ws client instance unique id. To not listen the echo to yourself.

players_ws_uid: String

json of vector of players for the server to know whom to send msg

players: String

json of vector of players with nicknames and order data

card_grid_data: String

vector of cards status

game_config: String

json of game_config

game_name: String

game name

MsgClick1stCard

player click

Fields of MsgClick1stCard

my_ws_uid: usize

this identifies the smartphone, but not the player-in-turn

players_ws_uid: String

all players for the server to know whom to send msg

card_index_of_first_click: usize

have to send all the state of the game

msg_id: usize

msg id (random)

MsgClick2ndCard

player click success

Fields of MsgClick2ndCard

my_ws_uid: usize

this identifies the smartphone, but not the player-in-turn

players_ws_uid: String

all players for the server to know whom to send msg

card_index_of_second_click: usize

have to send all the state of the game

is_point: bool

is point

msg_id: usize

msg id (random)

MsgDrinkEnd

drink end

Fields of MsgDrinkEnd

my_ws_uid: usize

this identifies the smartphone, but not the player-in-turn

players_ws_uid: String

all players for the server to know whom to send msg

MsgGameOver

Game Over

Fields of MsgGameOver

my_ws_uid: usize

this identifies the smartphone, but not the player-in-turn

players_ws_uid: String

all players for the server to know whom to send msg

MsgPlayAgain

Play Again

Fields of MsgPlayAgain

my_ws_uid: usize

this identifies the smartphone, but not the player-in-turn

players_ws_uid: String

all players for the server to know whom to send msg

MsgTakeTurn

player change

Fields of MsgTakeTurn

my_ws_uid: usize

ws client instance unique id. To not listen the echo to yourself.

players_ws_uid: String

all players for the server to know whom to send msg

msg_id: usize

msg id (random)

MsgAck

acknowledge msg, that the receiver received the message

Fields of MsgAck

my_ws_uid: usize

msg sender uid

players_ws_uid: String

send msg to this players

msg_id: usize

msg id (random)

msg_ack_kind: MsgAckKind

kind of ack msg

MsgAskPlayer1ForResync

ask player1 for resync

Fields of MsgAskPlayer1ForResync

my_ws_uid: usize

msg sender uid

players_ws_uid: String

send msg to this players

MsgAllGameData

all game data

Fields of MsgAllGameData

my_ws_uid: usize

ws client instance unique id. To not listen the echo to yourself.

players_ws_uid: String

only the players that reconnected

players: String

json of vector of players with nicknames and order data

card_grid_data: String

vector of cards status

card_index_of_first_click: usize

have to send all the state of the game

card_index_of_second_click: usize

have to send all the state of the game

player_turn: usize

player turn

game_status: GameStatus

game status